low-code-modeler: update update-lcm.sh
this change makes update-lcm.sh insert the contents of head.html and
body.html into the head and the body of index.html respectively.
the contents of head.html and body.html contain the code required to
make the low-code-modeler work with qhana-ui and save and load diagrams using
the api provided by this qhana-plugin-runner plugin.